refactor: rearrange files and dirs, added some tests
