+++ src/dataBlock/utility/subHeaderUtils.ts
@@ -1,0 +1,17 @@
+
+/**
+ * Gets the Subfile flags.
+ *
+ * @param  flag First byte of the subheader.
+ * @return The parameters.
+ */
+export class SubFlagParameters {
+  public changed: boolean;
+  public noPeakTable: boolean;
+  public modifiedArithmetic: boolean;
+  constructor(flag: number) {
+    this.changed = (flag & 1) !== 0;
+    this.noPeakTable = (flag & 8) !== 0;
+    this.modifiedArithmetic = (flag & 128) !== 0;
+  }
+}
